Ændret fra short til int

+++ EKG_Project/Files/RaspberryPi/FakeECGSensor.cs
@@ -5,7 +5,7 @@ public class FakeECGSensor : IECGSensor
     private int _i = 0;
     private readonly Random _rand = new Random();
 
-    public short ReadRawSample()
+    public int ReadRawSample()
     {
         // Simulerer EKG-lignende bølgeform + tilfældig støj
         double v = 1000 * Math.Sin(_i * 0.02) + _rand.NextDouble() * 50 - 25;
